refactor: move create_node tests to a separate script, move tmp_project_dir fixture to conftest.py
